first draft of python package
