+++ python_dataset_creator/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+"""
+The `python_dataset_creator` package provides functionality for creating datasets in Python.
+
+Note:
+This package is still under development and may not have all the desired features. Please refer to the documentation for more information on each module and function.
+
+Version: 1.0.0
+"""
