+++ python_dataset_creator/logging.conf
@@ -1,5 +1,5 @@
 [loggers]
-keys=root,data_loader,dataset_doc,cell_feature_doc,feature_defs_doc,image_settings_doc,dataset_writer, user_input_handler
+keys=root,data_loader,dataset_doc,cell_feature_doc,feature_defs_doc,image_settings_doc,dataset_writer,user_input_handler
 
 [handlers]
 keys=consoleHandler
@@ -54,4 +54,4 @@ args=(sys.stdout,)
 
 [formatter_simpleFormatter]
 format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
-datefmt=%m/%d/%Y %I:%M:%S %p
+datefmt=
